+++ examples/engine/engine.go
@@ -1,8 +1,8 @@
 package engine_contract
 
 import (
-	"github.com/CityOfZion/neo-storm/interop/engine"
-	"github.com/CityOfZion/neo-storm/interop/runtime"
+	"github.com/CityOfZion/neo-go/pkg/interop/engine"
+	"github.com/CityOfZion/neo-go/pkg/interop/runtime"
 )
 
 func Main() bool {
