+++ examples/token/token.go
@@ -1,10 +1,10 @@
 package token_contract
 
 import (
-	"github.com/CityOfZion/neo-storm/examples/token/nep5"
+	"github.com/CityOfZion/neo-go/examples/token/nep5"
 
-	"github.com/CityOfZion/neo-storm/interop/storage"
-	"github.com/CityOfZion/neo-storm/interop/util"
+	"github.com/CityOfZion/neo-go/pkg/interop/storage"
+	"github.com/CityOfZion/neo-go/pkg/interop/util"
 )
 
 const (
