--- interop/crypto/crypto.go
@@ -1,24 +1,0 @@
-package crypto
-
-// Package crypto provides function signatures that can be used inside
-// smart contracts that are written in the neo-storm framework.
-
-// SHA1 computes the sha1 hash of b.
-func SHA1(b []byte) []byte {
-	return nil
-}
-
-// SHA256 computes the sha256 hash of b.
-func SHA256(b []byte) []byte {
-	return nil
-}
-
-// Hash160 computes the sha256 + ripemd160 of b.
-func Hash160(b []byte) []byte {
-	return nil
-}
-
-// Hash256 computes the sha256^2 hash of b.
-func Hash256(b []byte) []byte {
-	return nil
-}
