--- pkg/vm/api/crypto/util.go
@@ -1,21 +1,0 @@
-package crypto
-
-// SHA1 computes the sha1 hash of b.
-func SHA1(b []byte) []byte {
-	return nil
-}
-
-// SHA256 computes the sha256 hash of b.
-func SHA256(b []byte) []byte {
-	return nil
-}
-
-// Hash160 ..
-func Hash160(b []byte) []byte {
-	return nil
-}
-
-// Hash256 ..
-func Hash256(b []byte) []byte {
-	return nil
-}
