--- pkg/vm/api/types/block.go
@@ -1,9 +1,0 @@
-package types
-
-// Block represents a block in the blockchain.
-type Block struct{}
-
-// Index returns the height of the block.
-func (b Block) Index() int {
-	return 0
-}
