--- pkg/vm/api/util/util.go
@@ -1,6 +1,0 @@
-package util
-
-// FromAddress returns the underlying bytes from the given address string.
-func FromAddress(address string) []byte {
-	return nil
-}
